requirements.txt
psycopg2==2.7.4
django==2.0
ujson==1.35
djorm-ext-pgfulltext==0.10
django-filter==1.1.0
graphene-django==2.1.0
psycopg2-binary==2.7.5